requirements.txt
Flask
gunicorn
python-dotenv
spotipy
SQLAlchemy
lyricsgenius
tqdm
requests